[Refactor] #3346 X11 にだけ依存するコードをmain/ からmain-unix/ へ移動させた
