+++ src/main-x11.cpp
@@ -95,9 +95,9 @@
 #include "io/files-util.h"
 #include "locale/japanese.h"
 #include "locale/utf-8.h"
+#include "main-unix/x11-type-string.h"
 #include "main/sound-definitions-table.h"
 #include "main/sound-of-music.h"
-#include "main/x11-type-string.h"
 #include "system/angband.h"
 #include "system/system-variables.h"
 #include "term/gameterm.h"
